@@ -1,5 +1,7 @@
-﻿using System;
+﻿using Microsoft.AspNetCore.Http;
+using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
@@ -14,5 +16,9 @@ namespace SimpleBlog2.Models
         public int ArticlePhotoId { get; set; }
         public int ArticleId { get; set; }
         public string FileName { get; set; }
+
+        [NotMapped]
+        [DisplayName("Wybierz zdjęcie")]
+        public IFormFile PhotoFile { get; set; }
     }
 }
