@@ -3,16 +3,22 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using SimpleBlog2.Models;
+using SimpleBlog2.ViewModels;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.EntityFrameworkCore;
+using System.IO;
 
 namespace SimpleBlog2.Repositories
 {
     public class ArticlePhotoRepository : IArticlePhotoRepository
     {
         private readonly SimpleBlog2Context _context;
+        private readonly IWebHostEnvironment _webHostEnvironment;
 
-        public ArticlePhotoRepository(SimpleBlog2Context context)
+        public ArticlePhotoRepository(SimpleBlog2Context context, IWebHostEnvironment webHostEnvironment)
         {
             _context = context;
+            _webHostEnvironment = webHostEnvironment;
         }
 
         public ArticlePhotoModel Get(int articlePhotoId)
@@ -24,6 +30,26 @@ namespace SimpleBlog2.Repositories
         public void Add(ArticlePhotoModel articlePhoto)
         {
             _context.ArticlePhotos.Add(articlePhoto);
+            _context.SaveChanges();
+        }
+        public void Add(ArticleCreateViewModel viewModel)
+        {
+            viewModel.ArticlePhoto.ArticleId = viewModel.Article.ArticleId;
+            viewModel.ArticlePhoto.FileName = viewModel.ArticlePhoto.PhotoFile.FileName;
+
+            //Save image to /images/Articles/
+            string wwwRootPath = _webHostEnvironment.WebRootPath;
+            string fileName = Path.GetFileNameWithoutExtension(viewModel.ArticlePhoto.FileName);
+            string fileExtension = Path.GetExtension(viewModel.ArticlePhoto.FileName);
+            viewModel.ArticlePhoto.FileName = fileName = fileName + DateTime.Now.ToString("yymmssffff") + fileExtension;
+            string filePath = Path.Combine(wwwRootPath + "/images/Articles", fileName);
+            using(var fileStream = new FileStream(filePath, FileMode.Create))
+            {
+                viewModel.ArticlePhoto.PhotoFile.CopyToAsync(fileStream);
+            }
+
+            //Insert into DB
+            _context.Add(viewModel.ArticlePhoto);
             _context.SaveChanges();
         }
 
